# src/web/openapi/davetiye.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Davetiye.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5052/"
    }
  ],
  "paths": {
...
    "/api/v1/invitations": {
      "get": {
        "tags": [
          "InvitationDraftEndpoints"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDraftPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "InvitationDraftEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDraftDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invitations/trash": {
      "get": {
        "tags": [
          "InvitationTrashEndpoints"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationTrashPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/invitations/{invitationId}/trash": {
      "post": {
        "tags": [
          "InvitationTrashEndpoints"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationTrashRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationTrashItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/{invitationId}/restore": {
      "post": {
        "tags": [
          "InvitationTrashEndpoints"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationTrashRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationStatus"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInvitationDraftRequest": {
        "required": [
          "contentSchemaVersion",
          "content",
          "templateKey"
        ],
        "type": "object",
        "properties": {
          "contentSchemaVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "content": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DraftContentInput"
              }
            ]
          },
          "templateKey": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "DraftContactInput": {
        "required": [
          "name",
          "role",
          "phone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DraftContentInput": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": [
              "null",
              "string"
            ]
          },
          "headline": {
            "type": [
              "null",
              "string"
            ]
          },
          "hostNames": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "startsAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "timeZoneId": {
            "type": [
              "null",
              "string"
            ]
          },
          "venue": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DraftVenueInput"
              }
            ]
          },
          "programItems": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DraftProgramItemInput"
            }
          },
          "contacts": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DraftContactInput"
            }
          },
          "announcement": {
            "type": [
              "null",
              "string"
            ]
          },
          "faqs": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DraftFaqInput"
            }
          },
          "transportStops": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/DraftTransportStopInput"
            }
          }
        },
        "additionalProperties": false
      },
      "DraftFaqInput": {
        "required": [
          "question",
          "answer"
        ],
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DraftProgramItemInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "startsAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DraftTransportStopInput": {
        "required": [
          "name",
          "address",
          "departureTime"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": [
              "null",
              "string"
            ]
          },
          "departureTime": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
      "DraftVenueInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "address": {
            "type": [
              "null",
              "string"
            ]
          },
          "mapUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "InvitationDraftDetails": {
        "required": [
          "id",
          "templateKey",
          "rendererVersion",
          "createdAt",
          "invitationRevision",
          "contentSchemaVersion",
          "content",
          "updatedAt",
          "contentRevision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "rendererVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "invitationRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "contentSchemaVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "content": {
            "$ref": "#/components/schemas/DraftContentInput"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "contentRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "InvitationDraftPage": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationDraftSummary"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "InvitationDraftSummary": {
        "required": [
          "id",
          "templateKey",
          "rendererVersion",
          "createdAt",
          "updatedAt",
          "invitationRevision",
          "contentRevision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "rendererVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "invitationRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "contentRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "effectiveState": {
            "type": "string",
            "default": "Draft"
          },
          "headline": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "InvitationTrashItem": {
        "required": [
          "invitationId",
          "headline",
          "deletedAtUtc",
          "purgeAfterUtc",
          "expected"
        ],
        "type": "object",
        "properties": {
          "invitationId": {
            "type": "string",
            "format": "uuid"
          },
          "headline": {
            "type": [
              "null",
              "string"
            ]
          },
          "deletedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "purgeAfterUtc": {
            "type": "string",
            "format": "date-time"
          },
          "expected": {
            "$ref": "#/components/schemas/PublicationRevisions"
          }
        }
      },
      "InvitationTrashPage": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "serverNowUtc"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationTrashItem"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "serverNowUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InvitationTrashRequest": {
        "required": [
          "expected"
        ],
        "type": "object",
        "properties": {
          "expected": {
            "$ref": "#/components/schemas/PublicationRevisions"
          },
          "expectedRetentionDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PublicationEntitlementStatus": {
        "required": [
          "maxPublishDays",
          "maxActiveInvitations",
          "premiumTemplatesEnabled"
        ],
        "type": "object",
        "properties": {
          "maxPublishDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "maxActiveInvitations": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "premiumTemplatesEnabled": {
            "type": "boolean"
          }
        }
      },
      "PublicationGrantChoice": {
        "required": [
          "grantId",
          "label",
          "kind",
          "maxPublishDays",
          "maxActiveInvitations",
          "premiumTemplatesEnabled"
        ],
        "type": "object",
        "properties": {
          "grantId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "maxPublishDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "maxActiveInvitations": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "premiumTemplatesEnabled": {
            "type": "boolean"
          }
        }
      },
      "PublicationPreflightStatus": {
        "required": [
          "templateAvailable",
          "isPremiumTemplate",
          "missingRequiredFields",
          "missingRecommendedFields"
        ],
        "type": "object",
        "properties": {
          "templateAvailable": {
            "type": "boolean"
          },
          "isPremiumTemplate": {
            "type": "boolean"
          },
          "missingRequiredFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "missingRecommendedFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PublicationRevisions": {
        "required": [
          "invitationRevision",
          "workingContentRevision",
          "publishedContentRevision",
          "windowId",
          "windowRevision"
        ],
        "type": "object",
        "properties": {
          "invitationRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "workingContentRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "publishedContentRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "windowId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "windowRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "PublicationStatus": {
        "required": [
          "invitationId",
          "publicCode",
          "storedState",
          "effectiveState",
          "serverNowUtc",
          "timeZoneId",
          "expected",
          "currentWindow",
          "published",
          "hasPendingChanges",
          "allowedActions",
          "preflight",
          "grantChoices",
          "currentEntitlements"
        ],
        "type": "object",
        "properties": {
          "invitationId": {
            "type": "string",
            "format": "uuid"
          },
          "publicCode": {
            "type": "string"
          },
          "storedState": {
            "type": "string"
          },
          "effectiveState": {
            "type": "string"
          },
          "serverNowUtc": {
            "type": "string",
            "format": "date-time"
          },
          "timeZoneId": {
            "type": "string"
          },
          "expected": {
            "$ref": "#/components/schemas/PublicationRevisions"
          },
          "currentWindow": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PublicationWindowStatus"
              }
            ]
          },
          "published": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PublishedSnapshotStatus"
              }
            ]
          },
          "hasPendingChanges": {
            "type": "boolean"
          },
          "allowedActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preflight": {
            "$ref": "#/components/schemas/PublicationPreflightStatus"
          },
          "grantChoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationGrantChoice"
            }
          },
          "currentEntitlements": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PublicationEntitlementStatus"
              }
            ]
          },
          "trashRetentionDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PublicationWindowStatus": {
        "required": [
          "id",
          "grantId",
          "startsAtUtc",
          "endsAtUtc",
          "timeZoneId",
          "revision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "grantId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endsAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "timeZoneId": {
            "type": "string"
          },
          "revision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "PublishedSnapshotStatus": {
        "required": [
          "revision",
          "sourceWorkingRevision",
          "templateKey",
          "rendererVersion"
        ],
        "type": "object",
        "properties": {
          "revision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "sourceWorkingRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "templateKey": {
            "type": "string"
          },
          "rendererVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}